pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MinionsLLM"
version = "1.0.0"
description = "A framework for generating behavior trees from natural language"
authors = [
    {name = "Andres Garcia", email = "a.a.garciarincon@student.vu.nl"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pygame==2.5.2",
    "numpy<2.0.0",
    "pandas==2.2.2",
    "violet-simulator==0.2.4",
    "llama-cpp-python==0.3.9",
    "openai>=1.0.0",
    "rouge-score==0.1.2",
    "matplotlib==3.9.0",
    "seaborn==0.13.2",
    "scikit-learn==1.5.1",
    "collections2==0.3.0",
    "huggingface_hub>=0.19.0",
    "lancedb>=0.22.1",
    "python-dotenv>=1.1.0",
    "ollama>=0.5.1",
    "datasets==3.6.0"
]

[tool.setuptools]
packages = [
    "data_grammar", 
    "data_grammar.grammar_gen", 
    "data_grammar.dataset_generation",
    "control_layer",
    "control_layer.simulation",
    "control_layer.simulation.agents",
    "control_layer.simulation.envs",
    "parser_layer",
    "llm_layer"
] 
